+++ example-views/testing/app/default/_controllers.php
@@ -4,33 +4,9 @@ use yii\helpers\Inflector;
 
 $favourites = ($favourites) ?: [];
 
+// Note: requires `$controllers` variable during rendering...
+
 ?>
-
-<div class="row">
-
-    <?php foreach ($favourites as $i => $controller): ?>
-
-        <?php
-        $className = $modelNamespace . Inflector::id2camel($controller['name']);
-        $count     = class_exists($className) ? $className::find([])->asArray()->count() : '-';
-        ?>
-
-        <div class="col-xs-6 col-sm-4 col-lg-3">
-            <?=
-            insolita\wgadminlte\SmallBox::widget(
-                [
-                    'head'        => $count,
-                    'type'        => $controller['color'],
-                    'text'        => $controller['label'],
-                    'footer'      => 'Manage',
-                    'footer_link' => $controller['route'],
-                    'icon'        => 'fa fa-' . $controller['icon']
-                ]
-            );
-            ?>
-        </div>
-    <?php endforeach; ?>
-</div>
 
 <div class="row">
     <div class="col-sm-12">