+++ example-views/testing/app/layouts/main.php
@@ -4,7 +4,7 @@ use yii\helpers\Html;
 
 /* @var $this \yii\web\View */
 /* @var $content string */
-$this->title = $this->title . ' [Backend] ' . Yii::$app->params['appName'];
+$this->title = $this->title . ' [Backend]';
 dmstr\web\AdminLteAsset::register($this);
 ?>
 
@@ -29,24 +29,23 @@ dmstr\web\AdminLteAsset::register($this);
     <![endif]-->
 </head>
 
-<body class="skin-black">
+<body class="hold-transition skin-black sidebar-mini">
 <?php $this->beginBody() ?>
 
 <div class="wrapper">
 
     <header class="main-header">
         <!-- Logo -->
-        <a href="<?= \Yii::$app->homeUrl ?>" class="logo"><?= getenv('APP_NAME') ?></a>
+        <a href="<?= \Yii::$app->homeUrl ?>" class="logo"><?= getenv('APP_TITLE') ?></a>
         <!-- Header Navbar: style can be found in header.less -->
         <nav class="navbar navbar-static-top" role="navigation">
             <!-- Sidebar toggle button-->
-            <a href="#" class="sidebar-toggle" data-toggle="offcanvas" role="button">
+            <a href="#" class="sidebar-toggle" data-toggle="push-menu" role="button">
                 <span class="sr-only">Toggle navigation</span>
             </a>
 
             <div class="navbar-custom-menu">
                 <ul class="nav navbar-nav">
-                    <?php if (!\Yii::$app->user->isGuest): ?>
                         <!-- Messages: style can be found in dropdown.less-->
                         <li class="dropdown messages-menu">
                             <a href="#" class="dropdown-toggle" data-toggle="dropdown">
@@ -67,27 +66,42 @@ dmstr\web\AdminLteAsset::register($this);
                                 </li>
                             </ul>
                         </li>
+                        <li class="dropdown tasks-menu">
+                            <a href="#" class="dropdown-toggle" data-toggle="dropdown">
+                                <i class="fa fa-flag"></i>
+                                <span class="label label-default">n/a</span>
+                            </a>
+                            <ul class="dropdown-menu">
+                                <li class="header">Languages</li>
+                                <li>
+                                    <!-- inner menu: contains the actual data -->
+                                    <ul class="menu">
+                                        languages
+                                    </ul>
+                                </li>
+                            </ul>
+                        </li>
                         <!-- User Account: style can be found in dropdown.less -->
                         <li class="dropdown user user-menu">
                             <a href="#" class="dropdown-toggle" data-toggle="dropdown">
                                 <i class="glyphicon glyphicon-user"></i>
-                                <span><?= \Yii::$app->user->identity->username ?> <i class="caret"></i></span>
+                                <span>Username <i class="caret"></i></span>
                             </a>
                             <ul class="dropdown-menu">
                                 <!-- User image -->
                                 <li class="user-header bg-light-blue">
                                     <?php echo \cebe\gravatar\Gravatar::widget(
                                         [
-                                            'email'   => \Yii::$app->user->identity->profile->gravatar_email,
+                                            'email'   => 'username@example.com',
                                             'options' => [
-                                                'alt' => \Yii::$app->user->identity->username
+                                                'alt' => 'username'
                                             ],
                                             'size'    => 128
                                         ]
                                     ); ?>
                                     <p>
-                                        <?= \Yii::$app->user->identity->username ?>
-                                        <small><?= \Yii::$app->user->identity->email ?></small>
+                                        username
+                                        <small>username@example.com</small>
                                     </p>
                                 </li>
                                 <!-- Menu Footer-->
@@ -103,7 +117,6 @@ dmstr\web\AdminLteAsset::register($this);
                                 </li>
                             </ul>
                         </li>
-                    <?php endif; ?>
                 </ul>
             </div>
         </nav>